refactor(settings): extract Editor / Appearance / Web panels (#1600, part 4)
Completes the SettingsDialog panel split. The final four Workspace/Ingest
panels come out of the monolith:

- EditorSettings.svelte — Done-batched (bind:editor; parent persists in
  handleDone via onApplyEditor).
- WebSettings.svelte — Done-batched (bind:webEnabled / allowedDomainsText /
  blockedDomainsText; parent builds the LLMSettingsUpdate in handleDone).
- AppearanceSettings.svelte — full extraction. Live-apply: the panel owns
  theme / fontFamily / editorFontSize / zoomPercent state, applies each
  immediately via its own $effects + onApplyFontSize/onThemeChanged callbacks,
  so the redundant handleDone appearance lines are gone.

ai was already extracted (AiSettings.svelte). Each sibling carries its own
copy of the shared .field/.hint styles and a flex-column root to match the
parent .panel gap. No behavior change — Settings → Editor / Appearance / Web
apply exactly as before.
